moved restart stage code to mom_player
To be better in line with mom_restart, moved most of the code in mom_restart_stage to the TimerCommand_RestartStage member function in mom_player.

Changed TimerCommand_RestartStage to take in stage & track arguments. If player is on same stage & track and timer is running, then it does executes the old code (does not restart timer & uses center of teleport as the destination for teleporting)